Remove split_diff_for_storage
